Move check let to file
