+++ src/types/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod blk;
+pub mod check_let;
 pub mod type_checker;
